refactor(address-list): 위젯 별도 파일로 분리, 사용하지 않는 import 삭제
